split fuzzing tests into separate files
