+++ tests/fuzzing/Tests/ReadOnlySpan/Linq/Sum.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.Fuzzing.TestHelper;
+
+namespace SpanExtensions.Tests.Fuzzing
+{
+    public static partial class ReadOnlySpanLinqTests
+    {
+        public sealed class Sum
+        {
+            // todo
+        }
+    }
+}
