--- tests/fuzzing/Tests/ReadOnlySpan/ReadOnlySpanTests.Linq.cs
@@ -1,9 +1,0 @@
-using static SpanExtensions.Tests.Fuzzing.TestHelper;
-
-namespace SpanExtensions.Tests.Fuzzing
-{
-    public static class ReadOnlySpanLinqTests
-    {
-
-    }
-}
