+++ tests/fuzzing/Tests/ReadOnlySpan/String/Remove.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.Fuzzing.TestHelper;
+
+namespace SpanExtensions.Tests.Fuzzing
+{
+    public static partial class ReadOnlySpanStringTests
+    {
+        public sealed class Remove
+        {
+            // todo
+        }
+    }
+}
