+++ tests/fuzzing/Tests/Span/Linq/Sum.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.Fuzzing.TestHelper;
+
+namespace SpanExtensions.Tests.Fuzzing
+{
+    public static partial class SpanLinqTests
+    {
+        public sealed class Sum
+        {
+            // todo
+        }
+    }
+}
