--- tests/fuzzing/Tests/Span/SpanTests.Linq.cs
@@ -1,9 +1,0 @@
-using static SpanExtensions.Tests.Fuzzing.TestHelper;
-
-namespace SpanExtensions.Tests.Fuzzing
-{
-    public static class SpanLinqTests
-    {
-
-    }
-}
