+++ tests/fuzzing/Tests/Span/Split/All.cs
@@ -1,0 +1,14 @@
+﻿using static SpanExtensions.Tests.Fuzzing.TestHelper;
+
+namespace SpanExtensions.Tests.Fuzzing
+{
+    public static partial class SpanSplitTests
+    {
+        const int count = 250;
+        const int minValue = 0;
+        const int maxValue = 100;
+        const int length = 100;
+        static readonly IEnumerable<StringSplitOptions> stringSplitOptions = GetAllStringSplitOptions();
+        static readonly CountExceedingBehaviour[] countExceedingBehaviours = (CountExceedingBehaviour[])Enum.GetValues(typeof(CountExceedingBehaviour));
+    }
+}
