+++ tests/fuzzing/Tests/Span/String/Remove.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.Fuzzing.TestHelper;
+
+namespace SpanExtensions.Tests.Fuzzing
+{
+    public static partial class SpanStringTests
+    {
+        public sealed class Remove
+        {
+            // todo
+        }
+    }
+}
